LayoutTests/imported/w3c/web-platform-tests/service-workers/service-worker/unregister-then-register-new-script.https-expected.txt



FAIL Registering a new script URL while an unregistered registration is in use assert_unreached: unexpected rejection: assert_equals: before activated registration.installing expected "https://localhost:9443/service-workers/service-worker/resources/empty-worker.js?new" but got "https://localhost:9443/service-workers/service-worker/resources/empty-worker.js" Reached unreachable code
FAIL Registering a new script URL that 404s does not resurrect an unregistered registration assert_unreached: unexpected rejection: assert_unreached: register should reject the promise Reached unreachable code Reached unreachable code
PASS Registering a new script URL that fails to install does not resurrect an unregistered registration 
